facilitator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
web3>=7.6.0
pydantic>=2.6.0
python-dotenv>=1.0.0
